- 指し手を表現する Move.cs追加 - Model.Shogi.Utilクラスをpartialに変更。それぞれのファイルに個別に書く。 	- Util.cs削除 - 各クラスのToInt()はUInt32で統一する。

+++ MyShogi/Model/Shogi/sfen/SfenUtil.cs
@@ -32,7 +32,7 @@ namespace MyShogi.Model.Shogi
         /// </summary>
         public static string PieceToSfen(Piece p)
         {
-            var c = SfenPieceList[p.RawPieceType().ToInt()];
+            var c = SfenPieceList[(int)p.RawPieceType().ToInt()];
 
             // 後手であれば小文字
             if (p.PieceColor() == Color.WHITE)
